+++ asset/integration_test.go
@@ -8,6 +8,7 @@ import (
 	"testing"
 
 	bolt "go.etcd.io/bbolt"
+
 	"github.com/sensu/sensu-go/asset"
 	"github.com/sensu/sensu-go/types"
 )